+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added `create-nostalgie-app`, the easiest way to scaffold out a new Nostalgie project.
+  
+  With a simple `npx create-nostalgie-app my-app` you have a fully-functioning Nostalgie app in less than 30 seconds.
+
 ### Changed
 - Increased logging during build and forced clean exit upon completion.